requirements.txt
streamlit
python-dotenv
PyMuPDF
langchain
faiss-cpu
openai
langchain-community
tiktoken